Refactoring in rates-service: changed exception messaging, validation classes and etc
