+++ driver-service/src/main/java/com/example/driverservice/service/DriverService.java
@@ -3,6 +3,7 @@ package com.example.driverservice.service;
 import com.example.driverservice.dto.common.PageResponse;
 import com.example.driverservice.dto.driver.DriverRequest;
 import com.example.driverservice.dto.driver.DriverResponse;
+import com.example.driverservice.dto.kafkaevent.RateChangeEventResponse;
 import jakarta.validation.Valid;
 import jakarta.validation.constraints.Min;
 import jakarta.validation.constraints.Positive;
@@ -15,8 +16,10 @@ public interface DriverService {
 
     DriverResponse create(@Valid DriverRequest driverRequest);
 
-    DriverResponse update(@Valid DriverRequest driverRequest,
-                          @Positive(message = "{validate.method.parameter.id.negative}") Long id);
+    DriverResponse updateDriver(@Valid DriverRequest driverRequest,
+                                @Positive(message = "{validate.method.parameter.id.negative}") Long id);
+
+    void updateRate(RateChangeEventResponse event);
 
     void delete(@Positive(message = "{validate.method.parameter.id.negative}") Long id);
